requirements.txt
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
pyyaml>=6.0
pytest>=7.4.0
pytest-cov>=4.1.0
flask>=3.0.0
plotly>=5.17.0
matplotlib>=3.8.0
